requirements.txt
Django>=5.0,<6.0
Pillow>=10.0.0
qrcode>=7.4
djangorestframework>=3.15.0
drf-spectacular>=0.27.0